requirements.txt
APScheduler==3.10.4
blinker==1.7.0
chromedriver_autoinstaller==0.6.4
openai==1.30.1
pandas==2.2.2
Pillow==10.3.0
PyAutoGUI==0.9.54
pyperclip==1.8.2
pyshorteners==1.0.1
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.20.0
selenium_wire==5.1.0
tqdm==4.66.2